fetcher-api/requirements.txt
fastapi
python-multipart
email-validator
ujson
python-dotenv
uvicorn[standard]
yt-dlp
